Refactor click event listeners, alter button functionality & attributes
The click event listeners for calculator buttons were initially spread
all over the place in the code making it a mess.
Hence, the click event listeners are now defined under one hood for every
button using the switch construct.

Replaced the id attributes of all buttons with value attribute as numbers
and operators are not valid id values.
Hence, the value attribute is now accessed to print to the screen.

Defined an operate() function which will later carry out calculation of
two numbers at a time and update result to the screen.
For the purpose of calculation of only two numbers at once, an operatorCount
variable has been defined which calls the opertor function ones it reaches
the value of 2.

Created a dotUsed boolean variable which stores whether the dot has been used
or not. Whenever a dot is used, it is set to true so that dot cannot be used
again. It is set back to false when 'AC' or 'Backspace' is used to erase it.