Define operate() function for calculation
Created a global array containing a list of operators that are used.
The array is used to check whether the screen text content contains
any of the operators given in array to carry out the required tasks.

The operate function now takes the input and extracts the operators
and depending upon the operator calls the designated functions for
the calculation tasks.
The functions are yet to be declared and updated.

Note: In one of the past commits, the operate() function is said to
be 'defined' instead of 'declared.' Kindly overlook the typo.