Refactor add() & subtract() functions, fix use of decimal button
The operate() function creates an array of numbers (of type string)
and passes them on to calculation functions.
Hence add() and subtract() now use .reduce() method to perform
calculations and return the answer.

Functions power() and percent() have only been declared.

Initially the decimal button could only be used once. Hence it could not
be used again for the second number (after the operator).
Therefore, the dotUsed variable is now set to 'false' each time an
operator button or '=' button is clicked.