Define remaining calculation functions, fix with NaN edge case
Completed definition for remaining calculation functions like power()
and percent().

Initially, when the user input two operators in a row without any
numbers, it produced 'NaN' on screen.
The issue has been fixed by replacing the 'NaN' with '0' on screen
so that the user can continue with calculations without any issues.