Fix operatorCount arithmetic and placement
Initially operatorCount was called and decremented inside operate()
function as well as when '=' was clicked which led to bug after the
usage of '=' where the input took in two operators instead of one
(as the count reached zero after getting decremented in both operator()
function and event listener of '='), hence leading to calculation errors.

The issue existed because operatorCount wasn't decremented at the right
place leading to excessive decrementation.

Therefore, the operatorCount is now decremented only when operator button
is clicked (provided the count has reached 2), or when '=' is clicked.
The operatorCount is no longer decremented from operate() function.